+++ src/SME.SGP.Dados/Repositorios/RepositorioProcessoExecutando.cs
@@ -106,7 +106,7 @@ namespace SME.SGP.Dados
             if (entidade.Id > 0)
                 await database.Conexao.UpdateMappedAsync(entidade);
             else
-                entidade.Id = (long)(await database.Conexao.InsertMappedAsync(entidade));
+                entidade.Id = (await database.Conexao.InsertMappedAsync(entidade));
 
             return entidade.Id;
         }
